Update video controller and views for concatenation of recordings
If there are multiple recorded segments of the same event, we want to
concatenate them before transcoding so that there is a single
published file (per format) for viewers to download.

The schema is being changed to add a video_event table augmenting the
event table (and a video_event_view that joins them).  This holds the
file base name for each recorded event, and video_target_file now has
an FK for video_event rather than for video_event_recording.

Change the controller and views to work with this new schema.