Lock video events and recordings while under review
Recordings must already be claimed/locked before review and can later
be unclaimed/unlocked by the holder or an admin.

We need to do the same for events, so that transcoding for an event
with multiple recordings is not started before all recordings are
associated with it.

Provide multiple 'finish' buttons on the review form which will unlock
the event and/or recording as appropriate.  Also add 'unclaim' actions
to the event list.